Moved comment author and element into archive directory
