+++ yark/archiver/archive.py
@@ -12,7 +12,7 @@ from .reporter import Reporter
 from ..errors import ArchiveNotFoundException
 from ..logger import _err_msg
 from .video.video import Video, Videos
-from .video.comment_author import CommentAuthor
+from .comment_author import CommentAuthor
 from typing import Optional
 from .config import Config
 from .converter import Converter
