+++ yark/archiver/video/comments.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 import multiprocessing
 from functools import partial
 from typing import Optional, Any, TYPE_CHECKING
-from .comment_author import CommentAuthor
-from .element import Element
+from ..comment_author import CommentAuthor
+from ..element import Element
 import datetime
 
 if TYPE_CHECKING:
